Fixed a problem w/ state reordering. Fixed a problem when selecting greyed out controlling character
git-svn-id: https://open-delta.googlecode.com/svn/trunk@1161 dd429c29-c832-59a0-e864-46e68368aa0b

+++ delta-editor/src/main/java/au/org/ala/delta/editor/ui/ControlledByEditor.java
@@ -68,7 +68,7 @@ public class ControlledByEditor extends CharacterDepencencyEditor {
 				for (int index : indices) {
 					for (CharacterDependency controllingAttribute : _controllingAttributes) {
 						if (_allControllingAttributes.indexOf(controllingAttribute) == index) {
-							newSelection.remove(index);
+							newSelection.remove((Integer)index);
 						}
 					}
 				}
